% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{apply_qc}
\alias{apply_qc}
\title{Apply QC Filtering Bounds}
\usage{
apply_qc(x, p, apply = TRUE)
}
\arguments{
\item{x}{A Seurat object or a list of Seurat objects.}

\item{p}{A list of QC parameters (e.g., qc$pre or qc$post).}

\item{apply}{If TRUE, physically subset cells; if FALSE, only write metadata flag.}
}
\description{
Apply QC Filtering Bounds
}
