% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{clean_header}
\alias{clean_header}
\title{Strip XML fragments / whitespace artifacts from a header string}
\usage{
clean_header(x)
}
\arguments{
\item{x}{Character vector of column headers.}
}
\value{
Cleaned character vector.
}
\description{
Strip XML fragments / whitespace artifacts from a header string
}
