% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_review.R
\name{cluster_review_params}
\alias{cluster_review_params}
\title{Default cluster-review parameters}
\usage{
cluster_review_params(...)
}
\arguments{
\item{...}{Named overrides.}
}
\value{
A named list.
}
\description{
Default cluster-review parameters
}
