% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge a sample list into one object}
\usage{
merge_samples(x, add_cell_ids = TRUE, project = "TamuScDSC")
}
\arguments{
\item{x}{A sample list (or, passed through, a Seurat object).}

\item{add_cell_ids}{Prefix cell names with the sample name to guarantee
unique barcodes (default TRUE).}

\item{project}{Project name for the merged object.}
}
\value{
A single Seurat object.
}
\description{
Merge a sample list into one object
}
