% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{norm_key}
\alias{norm_key}
\title{Normalize a key for tolerant matching (lowercase, alphanumeric only)}
\usage{
norm_key(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Normalized character vector.
}
\description{
Normalize a key for tolerant matching (lowercase, alphanumeric only)
}
